This argsort implementation also keeps track of the indexes
during the sort procedure. This remains cache efficient because
compared/swapped elements are always close. It still needs allocation
of double T and Idx buffers, but it saves one fine allocation that
was needed to unzip (T, Idx) as the two are now already separated.

Also a few improvements in default sort and vec creation (e.g. memcpy).


```
Gnuplot not found, using plotters backend
bench argsort null rate: 0%                                                                            
                        time:   [9.6161 ms 9.6518 ms 9.6886 ms]
                        change: [-9.5731% -8.2806% -7.0734%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild

bench argsort null rate: 20%                                                                            
                        time:   [16.184 ms 16.248 ms 16.316 ms]
                        change: [-28.536% -28.124% -27.757%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  8 (8.00%) high mild

```